refactor: split drawdown episodes
